requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-core
langchain-classic
sentence-transformers
faiss-cpu
pypdf
python-dotenv
huggingface-hub